# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden: <ul>
      <li><font color="red"><b>Wichtiger allgemeiner Hinweis:</b> Beachten Sie unbedingt die Benennung der Form-Parameter, Query-Parameter sowie die der Ergebnistupel. Bei Abweichungen von in der API angegebenen Benennungen wird der entsprechende Endpunkt als <u><b>falsch</b></u> bewertet.</font></li>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
servers:
...
paths:
...
  /mannschaften:
    get:
      tags:
         - Anwender
      summary: Gibt Mannschaften zurück.
      description: 
        Alle **Mannschaften** werden zurückgegeben. Falls Query-Parameter mit Werten angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: name
          in: query
          description: 'Nur Mannschaften, deren **Name** `name` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        mannschaftid:
                          type: integer
                          description: '**RowID** der Mannschaft.'
                      required:
                        - mannschaftid
                    - $ref: '#/components/schemas/mannschaft'
...
  /ligen/{ligaid}/spiele:
    get:
      tags:
        - Anwender
      summary: Gibt Spiele einer Liga zurück.
      description: 
        Alle **Spiele** einer **Liga** werden zurückgegeben.
      parameters:
        - name: ligaid
          in: path
          description: '**RowID** der Liga.'
          schema:
           type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        spielid:
                          type: integer
                          description: '**RowID** des Spiels.'
                      required:
                        - verletzungid
                    - $ref: '#/components/schemas/spiel'
        '404':
          $ref: '#/components/responses/404'
...
  /fans/{fanid}/mannschaft:
    get:
      tags:
        - Anwender
      summary: Gibt die vom Fan gefolgte Mannschaft zurück.
      description: 
        Die vom Fan **gefolgte Mannschaft** wird zurückgegeben.
      parameters:
        - name: fanid
          in: path
          description: '**RowID** des Fans.'
          schema:
           type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        mannschaftid:
                          type: integer
                          description: '**RowID** der Mannschaft.'
                      required:
                        - mannschaftid
        '404':
          $ref: '#/components/responses/404'
    post:
      tags:
        - Authentifizierter und als Fan authorisierter Anwender
      summary: Eine Mannschaft wird von einem Fan gefolgt.
      security:
        - basicAuth: []
      description: Die gewählte Mannschaft mit **MannschaftID** wird von dem authentifizierten Fan mit **FanID** gefolgt.
      parameters:
      - name: fanid
        in: path
        description: '**RowID** des Fans.'
        schema:
          type: integer
        required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    mannschaftid:
                      type: integer
                      description: '**RowID** der Mannschaft.'
                  required:
                    - mannschaftid
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    patch:
      tags:
        - Authentifizierter und als Fan authorisierter Anwender
      security:
        - basicAuth: []
      summary: Ändert die gefolgte Mannschaft.
      description: 
        Die **gefolgte Mannschaft** wird durch eine **neue Mannschaft** ersetzt.
      parameters:
        - name: fanid
          in: path
          description: '**RowID** des Fans.'
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    mannschaftid:
                      type: integer
                      description: '**RowID** der neuen Mannschaft.'
                  required:
                    - mannschaftid
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource. Siehe hierzu auch die Anmerkung bzgl. fiktiver Links im ersten Abschnitt.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
...
    spiel:
      type: object
      properties:
        link:
          type: string
          description: '**Link** des Spiels.'
        datum:
          type: string
          format: date
          description: '**Datum** des Spiels (in der Form "YYYY-MM-DD").'
      required:
        - link
        - datum
...
    mannschaft:
      type: object
      properties:
        name:
          type: string
          description: '**Name** der Mannschaft.'
      required:
        - name
...